bounding box + monster ia

+++ srcs/game/Monster.cpp
@@ -5,7 +5,7 @@
 // Login   <lafont_g@epitech.net>
 //
 // Started on  Sat May 12 09:47:20 2012 geoffroy lafontaine
-// Last update Mon May 21 15:26:48 2012 thibault carpentier
+// Last update Mon May 21 17:23:41 2012 Jonathan Machado
 // Last update Mon May 21 12:10:15 2012 lois burg
 //
 
@@ -65,8 +65,8 @@ void		Monster::update(gdl::GameClock& clock, gdl::Input& keys, std::list<AObject
 {
   brainScript_->updateDangerMap(objs);
   brainScript_->selectFct("thinking");
-  brainScript_->addParam(static_cast<int>(getPos().x));
-  brainScript_->addParam(static_cast<int>(getPos().y));
+  brainScript_->addParam(static_cast<float>(getPos().x));
+  brainScript_->addParam(static_cast<float>(getPos().y));
   brainScript_->callFct(1);
   update(clock, brainScript_->getDecision(), objs);
   (void)keys;
